+++ src/test/java/com/suresh/sms/dto/StudentDTOTest.java
@@ -1,5 +1,6 @@
 package com.suresh.sms.dto;
 
+import java.math.BigDecimal;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -21,14 +22,14 @@ class StudentDTOTest {
                 "Suresh",
                 "suresh@gmail.com",
                 "MCA",
-                50000.0
+                BigDecimal.valueOf(50000.0)
         );
 
         assertEquals(1L, student.getId());
         assertEquals("Suresh", student.getName());
         assertEquals("suresh@gmail.com", student.getEmail());
         assertEquals("MCA", student.getCourse());
-        assertEquals(50000.0, student.getFee());
+        assertEquals(BigDecimal.valueOf(50000.0), student.getFee());
     }
 
     @Test
@@ -39,12 +40,12 @@ class StudentDTOTest {
         student.setName("Ravi");
         student.setEmail("ravi@gmail.com");
         student.setCourse("Java");
-        student.setFee(30000.0);
+        student.setFee(BigDecimal.valueOf(30000.0));
 
         assertEquals(2L, student.getId());
         assertEquals("Ravi", student.getName());
         assertEquals("ravi@gmail.com", student.getEmail());
         assertEquals("Java", student.getCourse());
-        assertEquals(30000.0, student.getFee());
+        assertEquals(BigDecimal.valueOf(30000.0), student.getFee());
     }
 }
